Feature/fix ros2 UI (#148)
Description
This PR updates the web-ui to be compatible with ROS2
Changes include

Updating message types to the ROS2 versions
Updating the roslibjs versions to master commit 85fd9a7d817e77a030d594ac56b6c2b65f17659f (1.3 i think) which is also supporting ros2
Updated IsROSBridgeConnected() method to actually return connection status. Previously on every new message and additional listener was added to the connection emitter which is a memory leak and not at all needed. Now a single listener is created when connection to ros is established.
Added a subscription check for the call to checkSystemAlerts() to prevent repeated creation and deletion of the subscription at startup.
Related Issue
Supports usdot-fhwa-stol/carma-platform#1896

Motivation and Context