Car 6113 UI Fixes (#179)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
This PR fixes the steering and GPS visualizations. 

The steering visualization did not work because it needs a value for the
vehicle's steering ratio, which is available in
`VehicleConfigParams.yaml` in `carma-config`. This used to be made
available via the global parameter server in ROS 1, but since there is
no node that uses this parameter in ROS 2 the parameter is no longer
available. The solution was to pull the vehicle parameters directly from
the yaml file and load them to the UI.

The GPS visualization used the topic
`/hardware_interface/driver_status`, which is no longer used. The
visualization was updated to turn green if the vehicle is receiving GNSS
messages and turn grey if it has been longer than 0.5 seconds since the
last received message. Eventually we will want to update this logic to
include the other possible GPS states that were originally available.

An additional fix was needed on the Blue Lexus since the topic names are
slightly different:
https://github.com/usdot-fhwa-stol/carma-ssc-interface-wrapper/pull/179

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

<!-- e.g. CAR-123 -->
[CAR-6113](https://usdot-carma.atlassian.net/browse/CAR-6113)

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Needed for demos and general testing on the vehicle

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Confirmed that all of the above components worked on the White Pacifica
and the Blue Lexus

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[CAR-6113]:
https://usdot-carma.atlassian.net/browse/CAR-6113?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ