Update base image tag to 8.2-apache-bullseye (#185)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
Debian Buster has been EOL since 2022 and is no longer available
(causing the carma-web-ui Docker image build to fail), so the base image
is being updated from php:7.1-apache-buster to php:8.2-apache-bullseye.

While Bullseye is a newer Debian release than Buster, Bullseye is EOL as
well. However, updating to the latest supported Debian release
(Bookworm) causes a conflict with a GID 999 used by the CARMA system for
docker. Due to time constraints with current work, this will be resolved
at a future time, and for now this PR will only update to a newer base
image that is available.

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->
N/A

## Related Jira Key

<!-- e.g. CAR-123 -->
CDAD-173

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Debian Buster has been EOL since 2022 and is no longer available, which
causes the carma-web-ui Docker image build to fail.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
CI

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.